Wired up data, fixed persistence & UI, added tests
- Wired up the list of rooms with the viewcontroller
- Fixed persistence issues related to Resource loading
- Fixed NSInteger 64-bit compilation warnings
- Fixed persistence issue where resourceDescription is absent on (most)
  cases.
- Added 'reset' method on CABLResource to clear out cache.
- UI tweak on clock; fonts still not quite right.

+++ CalBrowser/SelectionViewController.m
@@ -72,14 +72,14 @@
     NSString *nextTimeBlock;
     
     if (components.minute <= 25) {
-        currentTimeBlock = [NSString stringWithFormat:@"%i:00", hour];
-        nextTimeBlock    = [NSString stringWithFormat:@"%i:30", hour];
+        currentTimeBlock = [NSString stringWithFormat:@"%i:00", (int)hour];
+        nextTimeBlock    = [NSString stringWithFormat:@"%i:30", (int)hour];
     } else if (components.minute <= 55) {
-        currentTimeBlock = [NSString stringWithFormat:@"%i:30", hour];
-        nextTimeBlock    = [NSString stringWithFormat:@"%i:00", (hour + 1) % 12];
+        currentTimeBlock = [NSString stringWithFormat:@"%i:30", (int)hour];
+        nextTimeBlock    = [NSString stringWithFormat:@"%i:00", (int)(hour + 1) % 12];
     } else {
-        currentTimeBlock = [NSString stringWithFormat:@"%i:00", (hour + 1) % 12];
-        nextTimeBlock    = [NSString stringWithFormat:@"%i:30", (hour + 1) % 12];
+        currentTimeBlock = [NSString stringWithFormat:@"%i:00", (int)(hour + 1) % 12];
+        nextTimeBlock    = [NSString stringWithFormat:@"%i:30", (int)(hour + 1) % 12];
     }
     
     //
